move AppNotification and AppNotificationSeverity in grafana-data
