+++ public/app/features/alerting/AlertTab.tsx
@@ -1,13 +1,13 @@
 import React, { PureComponent } from 'react';
 import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 
-import { EventBusSrv } from '@grafana/data';
+import { EventBusSrv, AppNotificationSeverity } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
 import { AngularComponent, config, getAngularLoader, getDataSourceSrv } from '@grafana/runtime';
 import { Alert, Button, ConfirmModal, Container, CustomScrollbar, HorizontalGroup, Modal } from '@grafana/ui';
 import EmptyListCTA from 'app/core/components/EmptyListCTA/EmptyListCTA';
 import { getPanelStateForModel } from 'app/features/panel/state/selectors';
-import { AppNotificationSeverity, StoreState } from 'app/types';
+import { StoreState } from 'app/types';
 
 import { AlertState } from '../../plugins/datasource/alertmanager/types';
 import { PanelNotSupported } from '../dashboard/components/PanelEditor/PanelNotSupported';
