+++ public/app/features/panel/components/PanelPluginError.tsx
@@ -2,9 +2,8 @@
 import React, { PureComponent, ReactNode } from 'react';
 
 // Types
-import { PanelProps, PanelPlugin, PluginType, PanelPluginMeta } from '@grafana/data';
+import { PanelProps, PanelPlugin, PluginType, PanelPluginMeta, AppNotificationSeverity } from '@grafana/data';
 import { Alert } from '@grafana/ui';
-import { AppNotificationSeverity } from 'app/types';
 
 interface Props {
   title: string;
