+++ public/app/types/appNotifications.ts
@@ -1,21 +1,4 @@
-export interface AppNotification {
-  id: string;
-  severity: AppNotificationSeverity;
-  icon: string;
-  title: string;
-  text: string;
-  traceId?: string;
-  component?: React.ReactElement;
-  showing: boolean;
-  timestamp: number;
-}
-
-export enum AppNotificationSeverity {
-  Success = 'success',
-  Warning = 'warning',
-  Error = 'error',
-  Info = 'info',
-}
+import { AppNotification, AppNotificationSeverity } from '@grafana/data';
 
 export enum AppNotificationTimeout {
   Success = 3000,
